Unifying cloud and atmosphere materials into a runtime RuntimeMaterialBinding (build-time custom-shader materials render as the error shader here)
